Basic login and logout functionality now working.
When the user logs in, they should be returned to the page that they
were on when they initiated the login. A few settings ensures that this
works correctly even for the root homepage.

Note that visitors won't need to be logged in to see everything, only to
make changes to things.