Added url and view for user profile editing.
This is a very simple form, currently only showing the username
(read-only), first name, last name, and email. At some point we will
want an avatar (probably just use gravatar for this) and a few custom
fields (e.g. organization).

Note that you must be logged in to edit your profile, hence the
@login_required decorator.