Added gravatar support; now showing in view and edit profile templates.
Gravatar is a web service providing cross-site avatar and profile
information, which is handy for our purposes with the RMG website.
Gravatar is especially nice because it will provide a neutral yet
distinguishing icon even for emails that don't have a Gravatar.

The user must have a (free) account at Gravatar to upload a custom
avatar; this is not managed at our website in any way.