Added code for loading thermo database when web server is started.
We will be using the same classes that work with the database in RMG Py
to work with them for the website, to avoid code duplication. This means
we won't be using Django models (though we will use forms) for this. The
database is to be stored in the database/ directory, which is not to be
included in this repository (it should have its own).