Added view for browsing the thermo database components.
The thermo database consists of a depository, a set of libraries, and a
set of group additivity values and corrections. With this view you can
move up and down in the hierarchy of the thermo database components, and
the page will present a sensible set of links for the level you are at.
The navigation bar at the top tells you where you are and allows for
backtracking up one or more levels.