Added capacity for dynamically drawing molecules on web server.
The key here is to use an <img> tag with the src pointing to an URL
that resolves to the molecule drawing view and contains the adjacency
list of the molecule we are trying to draw. The molecule_draw extension
in rmgpy.chem.ext is used to render the molecule. This approach is
particularly nice because it avoids having to muck around with temporary
files.

Note that our adjacency list format contains the newline character,
which can't be used in URLs. Semicolons are to be used instead.