Added jQuery and jsMath scripts to base template head.
The thinking is that these scripts are going to be used rather
extensively throughout the website, so we may as well make them
available everywhere. Since we'll be serving them from our server, there
isn't much cost to this, especially since they'll be cached on the end
user's computer after the first download.

Note that we will *not* be storing jQuery or jsMath in the repository.
You should obtain them yourself and place them in the rmgweb/media
folder, so that the web server can find them and git can ignore them.