Now drawing MoleculePattern objects using pydot.
pydot is a Python wrapper for the graphviz tools, which provide pretty
rendering of graphs. Viewing the adjacency list of MoleculePatterns in a
graphviz-generated image is much more appealing and useful than staring
at the adjacency list itself. The procedure is similar as for drawing of
Molecules: use an <img> tag with the appropriate src URL, and the server
will return a HttpResponse with a mimetype of image/png and data
provided from pydot.

This commit introduces a dependency on pydot.