Added handling of case when entry reference begins with initials.
The reStructuredText filter wants to treat these as ordinals in an
outline, which makes the resulting output look funny. The fix is to add
a backslash-space after the (first) initial, so that this behavior is
suppressed. Since we don't want people to be storing the reference in
the database with this unnecessary backslash-space, we must handle it
within the web server.