Moved SECRET_KEY setting to standalone module.
For security reasons, you should not commit the secretkey.py module to
the repository, and you should use your own unique value of SECRET_KEY
that you don't share with others. An example secretkey.py file has been
added, but do NOT use that key. An easy way to generate a key is to
generate a new Django project and take the key that it generates.