Wrapped thermo database loading inside a module-level function.
The idea here is to only load the thermo database once when the server
is started. Before we did this at module import; however, Django imports
the module multiple times, which is wasteful since the database loading
is slow. By wrapping in a function, we can execute the function body
only at certain times, and also test that the database isn't already
loaded. Now the thermo database is only loaded once per server restart,
as expected.