Created partial templates for various kinetics models.
This commit is parallel to the previous commit, where we did the same
thing for thermo models. Between the added comments and the splitting of
functionality, things are looking much cleaner than before.

Additionally, I've refactored things so that every kinetics parameter is
contained within the kineticsData variable. This means we don't need the
efficiencies, Trange, and Prange variables any longer. I've also placed
the dataFormat value as the first item in the kineticsData list, so that
we don't need the dataFormat variable either.

These changes are valuable today as improvements to readability, and
they will pay further dividends later when we reuse large bits of this
functionality.