Added view for generating thermo data for individual species.
This view is designed to show the results of searching across the entire
thermo database for the thermodynamics of a given species, including
estimating via group additivity. For entries that come from the
depository or libraries, a hyperlink to the full entry is provided. Note
that we have reused the thermo model templates, making this a fairly
straightforward commit.