Added form for specifying molecule to search for thermodynamics.
Currently this is a simple text area for you to fill in an adjacency
list. In the future we would rather use a molecule drawing toolkit so
the end user doesn't have to work with adjacency lists. Providing a
valid adjacency list will bring you to the thermo data view added in the
previous commit.