Converted thermo javascript functions into javascript classes.
This lets us take advantage of object-oriented programming to make the
code easier to read and more compact, and to avoid polluting the global
namespace with lots of functions. The other advantage is that the
javascript classes are pretty similar to the Python classes, so knowing
how to use one is knowing how to use the other.

I'm not too concerned about optimization of the javascript at this
point. Browsers these days are expending a lot of effort towards getting
blazing fast javascript engines, and we aren't really using tons of
javascript in the website yet.