Added plots to thermo data view for comparing found thermo models.
The plotting functions in thermo.js needed some tweaking to enable
plotting of multiple data series. This affects both the thermoData and
thermoEntry views. These can handle both single series and multiple
series, only showing a legend in the latter case.

The new plots can be slow to draw on the thermo data page due to the
large amount of javascript happenings (both jsMath and Highcharts), so
I've turned off animation to see if that helps.