Added support for handling KineticsDataModel objects in database.
This requires a template for showing the k(T) vs. T data points, a port
of the Python class to Javascript (in kinetics.js), adding handler code
to several views, and being able to use Django's templates to write
Javascript code for creating a KineticsDataModel object. All of this
seems to be working for my test case.