Added first version of reaction kinetics search form.
This is designed such that the user provides either one or two
reactants, and the view shows all reactions in all databases that
involve the given reactant(s). In particular, this approach does not
require the user to label the atom centers, so we can later use a
structure drawing applet instead of text adjacency lists. I'm not yet
sure how to handle the output from the kinetics groups.