In thermo and kinetics data views, plot heights adjust with number of lines.
When there are a large number of hits on the thermo or kinetics data for
a given species/reaction, the legend of the plots gets so large that it
cannot fit in the fixed volume provided previously. This commit makes a
crude attempt to correct for the legend size so that the plot is still
viewable. On web pages we have plenty of vertical space to work with, so
this isn't a problem.