Kinetics search now allows for specifying of products as well as reactants.
The reactants are required, while the products are optional. If products
are not given, all possible products will be returned. If products are
given, only those exact products will be returned.

A better regexer than I could probably combine the six expressions that
point to the kineticsData view into a single expression.