Restored loading of database components only when necessary for views.
The RMG database is large enough that it takes more than a few seconds
to load the entire database into memory each time I have to restart the
server. This isn't a problem for production purposes, but is a pain for
development. By loading components only when necessary, we may save
considerable idle time for developers.