Removed .html suffix from all URL patterns.
As was correctly pointed out, the pages being served are not real HTML
pages, but only have the HTML mimetype. (They are created dynamically by
Django in response to user actions.) Thus it is slightly misleading to
use the .html suffix, so I have removed it. Thanks to the nice Django
framework we can do this very easily by simply modifying the URL pattern
file urls.py.