Now showing database subsections in sorted order in thermo/kinetics views.
Since we store the depository, libraries, and groups in dictionaries,
the order of appearance on the thermo and kinetics views would otherwise
change every time the database is reloaded (e.g. when the server is
restarted). By sorting by label we make the appearance consistent and
easier to navigate. (Aside: should we be sorting by name instead?)