Updated kineticsData view to use new Reaction classes in rmgpy.data.kinetics.
In particular, after this commit the search functionality is able to
return reactions from the kinetics depository, libraries, and groups,
with kinetics for each. Note that the direction of the reaction is
currently the one for which kinetics is defined, *not* the one for which
the reactants -> products is always forward.