Updated import statements due to RMG-Py codebase reorganization.
The website uses the actual RMG-Py codebase to provide its
functionality, which lets us avoid code duplication. This means that the
website depends on the RMG-Py code. In that project I have recently
completed a large refactoring that moved a lot of modules around. This
commit updates this project accordingly, mostly in fixing of import
statements and changing from MoleculePattern to Group.