Added initial homepage for pdep app.
This page will probably act as the homepage for MEASURE. Thus it is
important that it look attractive, inviting, and (most of all)
functional. At the moment the page contains a banner, a (too brief)
description of MEASURE, and a big link to use to begin the process of
running a calculation.