Added form, view, and template for editing MEASURE input files in the browser.
This will be handy e.g. if the user wants to quickly change a small part
of the input file. Note that we will provide a separate means of simply
downloading the input file.

To implement this form, several methods were added to the Network class
involving loading and saving the MEASURE input file. A challenge as the
web interface develops will be to keep the various means of storing the
input file data in sync.