Added form, view, and template for uploading MEASURE input files.
This will be handy if the user already has an input file, e.g. if they
created it locally. One challenge with this functionality is that Django
generally tries to avoid deleting uploaded files, instead saving the
file with a slightly different name. This is not the behavior I want, so
I added functionality for deleting the existing input file right before
the new one is uploaded.