Improved display of file information on main network view.
The files now display with their size and modification time, along with
an icon and a download link. The file information and corresponding
actions are grouped by a <div> with a visual border and background. To
avoid repetition, a small template networkFile.html is used for each
file being displayed.

The icons used have been committed to the /media/icons directory.