Fixed URLs in database app to allow nested database subsections.
This allows us to group database components into folders on disk if
desired for organization. The challenge is that the forward slash, which
is used in RMG to indicate such a grouping, is also used as a delimiter
in most URLs. The answer is to craft a regular expression that matches
database components in groups, which is done in this commit.