Significant reorganization of the networkIndex view.
Previously this view was used as a place to collect links that invoke
various actions (e.g. running a job, generating a PES, etc.). This
functionality has been moved into the sidebar, which is finally getting
some use. Instead, this view now shows a summary of the information
contained in the network, and and provide links to see more details (e.g.
plots). When no input data is available (e.g. when the network has just
been created), the view adjusts to clearly indicate the various ways in
which input data can be provided. The view also encourages you to draw a
PES of your network, so that you can visualize your data and ensure that
it is correct.