Added warning to networkIndex view if PES is out of date.
If the user updates their input file, the PES is not automatically redrawn
because it can take some time depending on how big their network is.
Instead, we simply warn them that the image is probably out of date, and
invite them to regenerate the image whenever they wish.

In the future I expect to do a similar thing with the output data.