Added functionality for loading input/output files to Network model class.
The new Network.load() method will load the network's output file if it
exists, or the input file if not, and returned the loaded network as a
rmgpy.measure.network.Network object. This is done so we can use the
MEASURE classes and functions to browse the data, rather than needing to
construct loads of complicated Django models for the same purpose.
The parameters returned by the readFile() method are all stored as new
attributes of the Network model class, in case we want to use them later.