Added species, path reactions, and net reactions to networkIndex view.
In an earlier commit we moved the actions from the main part of this page
into the sidebar, precisely so that we could use the main part for this
information. Now, each species, path reaction, and (if MEASURE has been run)
net reaction is shown as a row in a table, along with a bit of metadata and
a link to see more details. All species are drawn where possible.