Moved common methods and templates from database app to main app.
The main app is now officially the base app on which all the other apps
build on top of. This makes it the place to store common widgets, fields,
models, views, templates, classes, functions, etc.

In this commit there are four common methods - getLaTeXScientificNotation(),
getStructureMarkup(), prepareThermoModel(), and prepareKineticsModel() -
and four common templates - thermoModel, thermoModelJS, kineticsModel, and
kineticsModelJS - that have been migrated out of the database app and into
the main app. A new module rmgweb.main.tools has been made for the four
common methods.