Added statmech.js Javascript file for molecular degrees of freedom.
This Javascript module is more or less a port of the rmgpy.statmech Python
module. It is needed for the same reasons that the thermo.js and kinetics.js
scripts are: to provide some client-side scripting (e.g. for plotting data).
All of the statistical mechanical models in rmgpy.statmech are represented
here, along with methods for computing thermodynamics (heat capacity,
enthalpy, and entropy), partition functions, and densities of states.