Added statesModel.html and statesModelJS.html template files.
These files take molecular degree of freedom parameters and produce HTML
and Javascript output based on their values. Like the corresponding thermo
and kinetics templates, these should be able to be used whenever states
data is to be rendered. Accordingly, they are placed alongside the thermo
and kinetics equivalents in the main/templates directory.