Added first version of networkSpecies view to pdep app.
This view shows lots of details for a given species in a reaction network.
If molecular degrees of freedom data is available, it is shown along with
plots of the partition function, density of states, and any hindered rotor
potentials. If thermo is available, it is shown along with plots of the
heat capacity, entropy, enthalpy, and Gibbs free energy. If collision
parameters are available, they are shown.

A bit of tweaking to the default CSS file was done to make the output
look a bit cleaner.