Now generating thermo data for plots in Python instead of Javascript.
I've decided that maintaining Javascript implementations of classes we
already have in Python is not worth the extra effort required to debug and
maintain them. As a result, we needed a way to generate the data used by
the thermo plots in Python so that we could pass them to the HTML and
Javascript templates. This commit adds that functionality.

In order to forbid people from trying to use the Javascript classes, I
have removed thermo.js from the repository. It's still available in the
history if anyone ever wants to see them. A few needed functions were
moved from thermo.js to thermoModel.js first.