Now generating kinetics data for plots in Python instead of Javascript.
Again, this is done to avoid needing Javascript versions of the kinetics
model classes. The challenge in generating kinetics data for plots is that
some kinetics are pressure-dependent. For now I am simply generating k(T,P)
at several pressures, so we can at least see some indication of the
pressure dependence. This will probably not play nicely with the kinetics
search functionality, but is okay for now.

In order to forbid people from trying to use the Javascript classes, I
have removed kinetics.js from the repository. It's still available in the
history if anyone ever wants to see them. A few needed functions were
moved from kinetics.js to kineticsModel.js first.