Now generating states data for plots in Python instead of Javascript.
Again, this is done to avoid needing Javascript versions of the states
model classes. In fact, it was when writing the statmech.js script that
I decided that this was not worth the time, especially when I'm soon
going to be wanting even more complex data being plotted. It also makes
sense to use the same methods to generate the data for the plots so that
you see exactly the same result as what is used when you run jobs.

In order to forbid people from trying to use the Javascript classes, I
have removed statmech.js from the repository. It's still available in the
history if anyone ever wants to see them. A few needed functions were
moved from statmech.js to statesModel.js first.