Added first version of networkPathReaction view to pdep app.
This view shows lots of details for a given path reaction in a reaction
network. If high-pressure limit kinetics is available, it is shown along
with an associated plot. If detailed transition state is available, it is
shown along with plots of the partition function, density of states, and
any hindered rotor potentials. In either case, a plot of the microcanonical
rate coefficient k(E) is always shown in one or both directions (whichever
is unimolecular).