Added first version of networkPlotKinetics view.
This view provides functionality for plotting k(T,P) values for all reactions
that start from the same reactant configuration on the same axes. I have
found these plots to be immensely useful in evaluating the behavior of the
reaction network as the temperature and pressure changes. A small form
allows you to manipulate the reactant configuration, the temperature used
for the k(T,P) vs. P plot, and the pressure used for the k(T,P) vs T plot.

A link to this view has been added to the output section of the sidebar.
(Is there a better place for it?)