Removed all references to an input wizard.
This feature has not yet been written, and if it is implemented in the
future, I don't think it will be done in the manner previously imagined.
The idea here is that the pdep app is now close enough to resembling
something usable that we shouldn't hesitate to make it available before
adding additional features.