Now using random hash as primary key for Network objects instead of integer.
The use of the hash makes it practically impossible for a user to guess the
identifier for a network that he or she does not have the exact URL to.
Since we allow anyone to edit the network if they know the URL, this
security through obscurity is very important. A 32-character hash should be
sufficiently obscure.