Some tweaks to the method of producing k(E) and rho(E) data.
The motivation here is to reuse some bits of this code to plot k(E) data
in a different view (that does not yet exist). To facilitate this, the
code that computes the k(E) -- and by extension rho(E) -- has been moved
into a standalone function so that it can be called in multiple places.

Also fixed the k(E) and rho(E) plots so that the do not use E = 0 as the
xmin.