Added features to draw_molecule page.  Now allows for Inchi and SMILES inputs as well as adjacency lists.
After submission of a single molecule identifier string, output fills in other forms of identifiers automatically.  (i.e. after typing in adjacency list, inchi and smiles will be outputted in the form as well as drawing of the structure.)

Fixed thermoData library search to reflect that the libraries are now stored as a dictionary instead of a list.