Added error message display to networkIndex view.
Invariably users will upload or edit an input file that is invalid for one
reason or another. There is a good deal of validity checking done in
MEASURE when the input file is loaded, with good error messages when one
or more invalidities are found. This commit enables these messages to be
shown (rather loudly, on purpose) to the user when the input file is
invalid, so that the user may fix these issues.