Added more stylish links to website components on RMG homepage.
Rather than place them in the sidebar, I've put them in the main part of
the page. The style is similar to the homepage of the Python documentation,
but adds icons. I made some very basic icons for the database and molecule
drawing apps; someone with more artistic talent is free to make better
ones in the future.