Updated thermoData.html template to properly use new thermo model templates.
Recently I changed how thermodynamics models are rendered in the HTML
templates and passed to Javascript for generating plots. However, it seems
I did not update the thermoData view (the result of a thermo search) to
reflect this. Note that the fixed code is actually shorter than before,
which was part of the motivation of switching approaches in the first
place.