First attempt at laying out kinetics groups in tree instead of table.
The tree better reflects the hierarchy in which the kinetics groups exist.
The implementation uses nested unordered lists, which makes sense for
hierarchical data such as this. The challenge, however, is to also show
some data for each group in something that resembles a table, but using
our nested lists. This is currently achieved using a floating div area;
is this the best CSS approach?

Note that the HTML for rendering the tree must be constructed in Python
since it requires recursive traversal of the database tree. The
getKineticsTreeHTML() method takes care of this for us.