Enabled collapsing and expanding of entries in tree in kinetics view.
An image now appears next to each entry in the tree. When the image is
clicked, the children of that entry are shown or hidden, and the image
updated to reflect that change in status. Some Javascript and jQuery
manipulation is used to accomplish this.

I've also added image files that indicate the status of the tree entry
(collapsed, expanded, or no children) to the /media folder.