Now showing the kinetics group corrections at several temperatures.
This provides a way to see all of the group values at once. By using
k(T) this approach is independent of the form that the group additivity
values take (Arrhenius, KineticsData, etc.) If we decide to stick with
Arrhenius, though, we would want to think about printing A, n, and Ea
here instead of k(T) values.