Updates to get the website working with new kinetics families.
This fixes the kinetics views so that they work with the new kinetics
database organization. One benefit of this is that all of the pieces
that correspond to a single reaction family are now visually grouped
together on the website (as well as physically grouped together on disk
and programmatically grouped together in code).

Most of the work involved removing references to the kinetics depository
and replacing the references to the groups with references to the combined
family (which is, roughly speaking, the union of the depository and the
groups).

A particular challenge with this change was in allowing multiple levels
in the subsection parameter for a kinetics family (previously they could
be handled for a kinetics library).