Better styling of login and edit profile forms.
Mostly this is placing these forms in a table in the corresponding
templates, adding class names for styling purposes, and defining some
style attributes in CSS. Note the neat way Django provides for iterating
over the fields of a form, which is used in both places to avoid duplicate
code.

I've also added a CSS file for use within the main app only.