Fixed edit profile form so that it actually saves its contents.
For this to work as intended, you actually need separate forms for the
User and UserProfile models, and to validate and save both of them when
processing the view. This actually makes the form classes simpler, since
they don't need to custom declare form fields for the items in the User
model (and shouldn't, as it wouldn't be easy to move that information to
the User class before saving).

Also added the DivErrorList class and an associated style for cleaner
error messages.