Added form used for signing up for a user account.
This form asks for a minimal amount of information required to create the
user account. In particular, this form asks for a username, and validates
it so that it only uses letters, numbers, and underscores and is not
already taken.

Note that nothing yet happens when a valid signup form is submitted, as
I'm not sure exactly how we should handle that yet.