Added list of networks to user profile view.
Now that we are identifying Network models with the User that created them,
this is very simple to do. Each network is displayed with it short hash,
title, and last modified date.

Note that anyone with the network URL can still modify the network; this
only provides users a neat way of seeing all the networks they've created
and navigating to them.