Moved RMG-Java kinetics querying to new database.tools module.
The RMG-Java kinetics querying works nicely, but it's significant enough
that it deserves to be outside of the kineticsData view. (This would also
allow us to potentially query it in other situations if so desired.) The
code was moved to a new module rmgweb.database.tools; otherwise it is
mostly intact as-is.