kineticsData view now always shows reactants on left of reaction arrow.
If the kinetics are defined in the other direction, the reverse kinetics
are generated and used. An asterisk appears to let the user know that
we have done this. Since we need thermo data to generate reverse kinetics,
the thermo database is now loaded as part of this view.

If the kinetics are defined in that direction, then nothing changes.