Added kineticsResults view as intermediate search results page.
This view acts as an intermediate between the search page and the
kineticsData view. It provides a summary of the different reactions found
(groups by reactants and products only, not distinguished by reaction
family), along with the number of matches for each hit. The user can then
click to visit the kineticsData view for each match separately.