Added reactionHasReactants() function.
This function checks whether a given set of species serves as the exact
set of reactants for a given reaction. At the moment we use it to determine
if a set of reactions have the specified species in the reactant side or
not, so that we can make this condition true if it is not already so.