Adjacency list view first tries using query string as SMILES.
First we try using the query as a SMILES string, and only if that fails
do we pass it through the NCI resolver to get a SMILES string. This 
will both speed up requests that are SMILES strings, and avoid the 
SMILES-to-SMILES bugs in the NCI resolver for radical species.
(NCI 'resolves' [CH3] into C)