Molecule search form uses AJAX to create Adjacency List from identifier.
Only the adjacency list part of the form is used for drawing the molecule
or searching the database. The InChI or SMILES reading is now done via AJAX
through the /adjacencylist/ view created in the previous commit.

The identifier is translated as soon as focus leaves it if it has changed.
The results are hidden as soon as you edit the identifier.
Pressing return in the identifier field does not submit the form but instead
resolves the identifier and selects the form submit button.