Kinetics search form has chemical identifier lookup.
This is similar to the one I just added to the molecule search page.
Now if you press 'enter' in a form field it jumps you to the submit button;
if you press 'enter' when you're already there, it submits the form. Hopefully
this gives the adjacency lists time to load.