Create reusable getReactionUrl(reaction) function.
This is so we can re-use it in other views.
To generalize I had to make it cope with both molecules and species 
- I'm not sure when or why we have both, but this works. 

Also extended the urls so the 'reverse' call copes with three reactants in a reaction
(I can't remember where this showed up, but it did. Probably in a reaction library).