Fix bug in RMG-Java kinetics search: matching reaction strings.
The previous scheme would return True for
 A + B --> C + D
when you searched for
 A + A --> C + D
becuse it would check that both A and A could be found in the reactants
(which is true) and both C and D can be found in the products (which is true).

The new scheme means they both have to match exactly.