Some sprucing up of the RMG homepage.
I've moved the links to the different parts of the website into the
sidebar in anticipation of using the main part of the window for
something similar to the current RMG homepage on Sourceforge. I have
already grabbed the flux diagram from the Sourceforge homepage and
put it on this new page.

I've also added a direct link to the kinetics search tool, which will
be of interest to many of our users.

Still some work to be done here, but it's a step in the right direction
I think.