Moved a few settings from rmgweb.settings to rmgweb.secretsettings.
These settings are those that will differ between different users'
development environments, as well as the production environment. In
particular, a few settings that users wouldn't want to be public
(e.g. usernames and passwords) are in the secretsettings module so
that they are not placed into the history.

The example file should have defaults that work for most developers in
their local environments.