Sidebar now contains links to website sections on all pages.
This is done by placing the HTML defining the links in the base.html
template, so that they always appear. You can now use the sidebar_items
block to add custom items to the sidebar; they will appear below the
logo but above these links.

The styles that define the sidebar links were moved to default.css so that
they would be available on every page.

Closes #12.