Added generateReactions() function to database tools.
This function collects the code for querying RMG-Py and RMG-Java for
reactions and associated kinetics, as we use this functionality in multiple
places (specifically kineticsData and kineticsResults). Before I had
updated only the former with a more thorough algorithm, leading to the
website displaying a different number of results in the two views. Now they
use the same code, and therefore generate the same number of results.

Closes #8.