Convert ArrheniusEP kinetics to Arrhenius in kinetics search results.
Since we are searching for real reactions, we have access to the reactants
and products, and therefore can compute the thermo needed for the enthalpy
of reaction, and can do this conversion. This allows for easy comparison to
the other search results, especially for cases where alpha is nonzero
(which is not very often in our current set of rules, but it happens once
in a while).

We also needed to do this because you can't generate the reverse of
ArrheniusEP kinetics, and the kinetics search would try to do so if the
rule was in the reverse direction of what you were searching for.