reactionMatchesReactants() now handles the A + A --> products checking.
Before this, you would get the A --> products reactions in the wrong
direction because it was checking for A + A due to the search for
A + A --> products reactions. Two changes were needed for this: (1) make
a new list of the bimolecular reactants A + A to avoid trampling over the
original, and (2) allow for the case of one reactant passed as a parameter
but two products in the reaction when using reactionMatchesReactants(). In
the latter we require that *both* products are the specified reactant in
order to return True.

Closes #16.