Updated Network model to use new MEASURE class.
The MEASURE class, recently added to rmgpy.measure.main, encapsulates
the parameters needed for running a pressure-dependence calculation. This
means we don't have to store all of them on the Network model any longer;
instead we can simply store the MEASURE object itself. Note that, at the
moment, the pdep app only really uses the Network object, although that
will likely change as we start to add more input functionality.