Pass reference object instead of string to reference template.
This means we don't also need to pass referenceLink, we just get 
it from reference.url in the template. I'm not sure what's going on with
referenceType, as it seems to be just set equal to the empty string. 
For now I've left this alone. Main reason for this change is that the template
will now be able to interrogate attributes of the reference, rather than
just having a string.