Added render_kinetics.py with render_kinetics_math custom template tag.
The render_kinetics_math template tag can be used to render a kinetics
model using jsMath. To use in a template, first {% load render_kinetics %}
at the top of the template, and then use the syntax

{{ kinetics|render_kinetics_math }}

to call the filter. This filter should provide a nice replacement for the
ugly prepareKineticsParameters() function.