Added separate getAllKinetics() step to generateReactions() for website.
In the upstream RMG-Py, we no longer return the kinetics for template
reactions during KineticsDatabase.generateReactions(), thus necessitating
this change. Unlike the similar change made for RMG jobs, here we want to
keep every matching kinetics, not just the best.

Note that I've commented out the passing of the degeneracy to the final
reaction, as I don't think this is done correctly at the moment.