Fix bug in database.tools.generateReactions() that was killing RMG-Java search.
The block of code introduced in 41bf45d8c6684bb682c60295d5a7999da4706167
used the variables 'reactants' and 'products' to store and shuffle some
Species objects. Unfortunately these variables were already in use to store
the Molecule objects that had been passed in to the function as parameters.
Later, when searching for RMG-Java reactions, we need the original passed in 
parameters.

Is this a weakness of python (no static typing) or our naming conventions
(should we prefix all parameters with 'p_' or something?) or is there some
other way to minimize this type of bug? (which took me a while to find)