Form to edit reactions in kinetics depositories.
Visit, for example /database/kinetics/families/H_Abstraction/training/302/edit
and you can edit the data.

The form is pre-filled with the current entry.
When you submit, it validates the data, adds the history event,
and either displays the raw text or renders it as a kinetics template.
The next step is to have it save...