Load, store, and report thermo libraries in a specified order.
When estimating kinetics, we have to calculate the enthalpy of reaction.
It therefore matters what order our thermo libraries are stored in, as we
search from first to last. It seems the load function, which just walks
through the folder, varies from computer to computer, and certainly doesn't
know which are best. We now load them in the order:
 'primaryThermoLibrary', 'DFT_QCI_thermo', 'GRI-Mech3.0',
 'CBS_QB3_1dHR', 'KlippensteinH2O2', everything else...

This is somewhat arbirtary, and currently hard-coded into the load function,
but at least somewhat sensible, and consistent.