New kineticsGroupEstimateEntry view to see an individual family group estimate.
This view lets you see an estimate from group additive kinetics
from an individual reaction in an individual family.
The url is based on the family. It re-uses the kineticsEntry template,
just like the other kineticsEntry views.

getReactionUrl() has been modified to take family parameter, which 
will create the url to this new view, rather than to the search 
resulsts page.