New kineticsEntryNewTraining view to make new training reaction entry.
This lets you make a new entry in a family's training depository.
Perhaps it should be generalized to allow new entries in any
section and subseciton of the database, but this is my first priority.
(I thought it was going to be more specialized when I started,
but I now think it could be generalized quite easily).