kineticsGroupEstimateEntry now lets you add to training depository.
The view of a group additive estimate reaction has a button to
let you add this to the training depository.
This button perhaps could/should be added to the other kineticsEntry
views, but at least having it on the estimate one means we can
add it for all discovered reactions, not just those in PrIMe.