Pass degeneracy through to reactions in generateReactions()
The commit message to 41bf45d8c6684bb682c60295d5a7999da4706167 says
that this was commented out because it's probably not right.
Without it though, the pre-filled form for adding new training reactions
always has degeneracy=1. I am not sure what is or might be wrong with this,
but we should look into it and make sure it's right.