Fix bug to re-enable RMG-Java kinetics search.
When generateReactions is called with the optional only_families parameter
it should not try to generate RMG-Java kinetics, instead only returning
kinetics from the requested RMG-Py templates. However, the if/else statements
were the wrong way around so it would *only* call RMG-Java in the cases
when it is not supposed to.