Reload parts of the database if they have changed on disk since last loaded.
This is important now that you can update the database online, as apache wsgi
keeps several threads alive serving requests, and so they can get out of sync.
Now, whenever loadDatabase(component,section) is called, it checks to see
if the files on disk have changed, and if so it reloads that part of the
database.

I think/hope the loadDatabase is called at the top of each view that needs it.

This will probably slow us down a little - but hopefully walking the file system
and checking timestamps is not too slow..