Move the database reloading stuff from views.py into tools.py
Hopefully this is a good idea. The views file is too big anyway, and this 
would also allow it to be split up into separate files that all import the
tools.

I also made the "Resetting 'last loaded' timestamps" message report the process id, 
which may be interesting when reading the Apache log.