getReactionURL(reaction) returns '' if it finds Group or LogicNode
Previously it would just throw an exception. The point of this is to find
a url to search for additional kinetics, so it only works for real Molecules
or Species.  The place where we check has been moved to inside the function,
and has been extended to check for LogicNodes.