Fixed bug in networkPathReaction view involving k(E) calculation.
Recent changes in RMG-Py have modified how k(E) is computed. The bug was
caused by the fact that we now need E0 for all configurations (including
product channels), a change that was made in RMG-Py but not in this
copy-pasted duplicate. Instead, I have replaced this with a call to
the new Network.calculateGroundStateEnergies() method.